refactor: scss to css, can now use composes
