+++ src/pages/resources.module.css
@@ -1,0 +1,60 @@
+.resources {
+  width: 100%;
+  display: grid;
+  grid-gap: 2rem;
+  grid-template-columns: var(--content-grid-columns);
+  grid-template-areas: var(--content-grid-template);
+}
+
+.sidebar div section {
+  grid-gap: 0.5rem;
+}
+
+.content {
+  grid-area: content;
+  width: var(--content-width);
+  max-width: var(--max-content-width);
+  margin: 0 auto;
+}
+
+.content > div {
+  width: 100%;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.content h1 {
+  font-size: 2rem;
+  line-height: 1.3;
+}
+
+.content ul li::marker {
+  color: var(--primary-accent);
+}
+
+.content li {
+  font-weight: 450;
+  line-height: 1.75;
+}
+
+@media (max-width: 767px) {
+  .content header + div {
+    width: 100vw;
+    position: relative;
+    left: 50%;
+    right: 50%;
+    margin-left: -50vw;
+    margin-right: -50vw;
+    border-radius: 0;
+  }
+
+  .content header + div img {
+    border-radius: 0;
+  }
+}
+
+.content p {
+  line-height: 1.75;
+  font-size: 1.125rem;
+  font-weight: 450;
+}
