--- src/templates/layout.module.scss
@@ -1,155 +1,0 @@
-.blogpost,
-.resources {
-  width: 100%;
-  display: grid;
-  grid-gap: 2rem;
-  grid-template-columns: var(--content-grid-columns);
-  grid-template-areas: var(--content-grid-template);
-}
-
-.blogpost {
-  @media (max-width: 767px) {
-    --content-grid-template: "content" "moreposts";
-  }
-}
-
-.content {
-  grid-area: content;
-  width: var(--content-width);
-  max-width: var(--max-content-width);
-  margin: 0 auto;
-
-  > div {
-    width: 100%;
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  h1 {
-    font-size: 2rem;
-    line-height: 1.3;
-  }
-
-  ul {
-    li::marker {
-      color: var(--primary-accent);
-    }
-  }
-
-  li {
-    font-weight: 450;
-    line-height: 1.75;
-  }
-
-  @media (max-width: 767px) {
-    header + div {
-      width: 100vw;
-      position: relative;
-      left: 50%;
-      right: 50%;
-      margin-left: -50vw;
-      margin-right: -50vw;
-      border-radius: 0;
-
-      img {
-        border-radius: 0;
-      }
-    }
-  }
-
-  p {
-    line-height: 1.75;
-    font-size: 1.125rem;
-    font-weight: 450;
-  }
-}
-
-.sidebar {
-  @media (max-width: 767px) {
-    display: none;
-  }
-}
-
-.resourceSidebar {
-  div {
-    section {
-      grid-gap: 0.5rem;
-    }
-  }
-}
-
-.bar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 1rem;
-}
-
-.posts {
-  display: grid;
-  grid-row-gap: 1rem;
-}
-
-nav {
-  padding-top: 2rem;
-  margin-bottom: 4rem;
-  width: 80%;
-  margin-left: auto;
-  margin-right: auto;
-
-  ul {
-    margin: 0;
-  }
-}
-
-.tableOfContents {
-  margin-bottom: 0;
-
-  a {
-    text-decoration: none;
-    color: var(--primary-accent);
-  }
-  ul {
-    list-style-type: none;
-    font-weight: 400;
-    p {
-      font-weight: 550;
-    }
-    li > ul {
-      margin-top: 0.6rem;
-    }
-    li {
-      margin-bottom: 0.6rem;
-    }
-  }
-  > ul {
-    margin: 0;
-  }
-}
-
-.tags {
-  margin: 0;
-
-  div {
-    cursor: pointer;
-    font-size: 0.825rem;
-    letter-spacing: 0.02rem;
-    background: var(--primary-white);
-    padding: 0.25rem 0.5rem;
-    display: inline-block;
-    margin-right: 0.5rem;
-    margin-bottom: 0.5rem;
-    color: var(--primary-accent);
-    text-transform: capitalize;
-    font-weight: 500;
-    border: 1px solid var(--primary-accent);
-    border-radius: 1px;
-  }
-}
-
-.written {
-  p:last-of-type {
-    margin-top: 0.2rem;
-    color: var(--primary-grey);
-  }
-}
